Changed some files to hidden

+++ .wpilib-download.sh
@@ -9,14 +9,22 @@
 
 # TODO: script determining the latest .jar file based on the date string
 # rather than a hard coded file name.
+version="0.1.0.201501221609"
 
-wget --quiet -O wpicpp.zip http://first.wpi.edu/FRC/roborio/release/eclipse/plugins/edu.wpi.first.wpilib.plugins.cpp_0.1.0.201501170231.jar 
-unzip wpicpp.zip resources/cpp.zip
-mkdir wpilib
-mv resources/cpp.zip ./
-rm -rf resources
-unzip -d wpilib/ cpp.zip
+source wpilib/version.txt
+if [ ! "$version" = "$downloaded_version" ] ; then 
+	wget --quiet -O wpicpp.zip http://first.wpi.edu/FRC/roborio/release/eclipse/plugins/edu.wpi.first.wpilib.plugins.cpp_$version.jar 
+	unzip wpicpp.zip resources/cpp.zip
+	mkdir wpilib
+	mv resources/cpp.zip ./
+	rm -rf resources
+	unzip -d wpilib/ cpp.zip
 
-rm -rf cpp.zip
-rm -rf wpicpp.zip
+	rm -rf cpp.zip
+	rm -rf wpicpp.zip
+	echo "downloaded_version=$version" > wpilib/version.txt
+else
+	echo "Already at latest version"
+fi
 
+echo "Version = $version"
